feat: implement search command for Jikan.js with character and anime subcommands
